fix(infrastructure): Resolve NATS type mismatch (u32→i64)
- Change NATS_CONSUMER_MAX_DELIVER constant from u32 to i64 to match field type
- Remove unnecessary type cast in nats.rs
- Compilation now clean (0 errors, pre-existing warnings remain)

Phase 0 pre-flight validation complete. Ready for Phase 1 domain refactoring.

Validation:
- cargo check --lib: ✓ PASSED
- Exit code: 0

+++ src/infrastructure/events/nats.rs
@@ -225,7 +225,7 @@ impl EventBusProvider for NatsEventBus {
                     deliver_policy: jetstream::consumer::DeliverPolicy::New,
                     ack_policy: jetstream::consumer::AckPolicy::Explicit,
                     ack_wait: NATS_CONSUMER_ACK_WAIT,
-                    max_deliver: NATS_CONSUMER_MAX_DELIVER as i64,
+                    max_deliver: NATS_CONSUMER_MAX_DELIVER,
                     ..Default::default()
                 },
                 &self.stream_name,
